requirements.txt
bcrypt==4.0.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
flask-msearch==0.2.9.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==3.0.1
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
SQLAlchemy==1.4.39
sqlparse==0.4.4
Werkzeug==2.2.0
WTForms==3.1.1
